Fix: Resolve merge conflict in core/extension/build.gradle.kts (#110)
* Initial plan

* Fix legacy extension compatibility: correct metadata keys, add SourceFactory, fix build deps


* Fix encodedSchemeSpecificPart -> schemeSpecificPart in ExtensionInstallReceiver


* Update core/extension/src/main/java/app/otakureader/core/extension/data/remote/ExtensionRemoteDataSource.kt


* Update core/tachiyomi-compat/src/main/java/eu/kanade/tachiyomi/source/Source.kt


* Update core/tachiyomi-compat/consumer-rules.pro


* fix: use eu.kanade.tachiyomi.source types in ExtensionLoader; per-broadcast scope in receiver


* fix: resolve merge conflict in core/extension/build.gradle.kts - add projects.sourceApi from main


* Merge main into legacy extension compatibility branch to resolve conflicts (#114)

* Implement Local Backup & Restore (#108)

* Initial plan

* Implement local backup and restore functionality

- Add backup data models with kotlinx.serialization
- Create BackupCreator to export library, categories, history, and preferences to JSON
- Create BackupRestorer to import data back into Room database
- Add BackupRepository to coordinate file I/O operations
- Update Settings screen with backup/restore UI and file pickers
- Add progress indicators and snackbar feedback
- Fix missing core:extension build.gradle.kts file


* Update data/src/main/java/app/otakureader/data/backup/repository/BackupRepository.kt


* Update data/src/main/java/app/otakureader/data/backup/BackupCreator.kt


* Update feature/settings/src/main/java/app/otakureader/feature/settings/SettingsScreen.kt


* Update data/src/main/java/app/otakureader/data/backup/BackupCreator.kt


---------


* Initial plan

---------


---------

+++ core/tachiyomi-compat/src/main/java/app/otakureader/core/tachiyomi/compat/TachiyomiManifestParser.kt
@@ -364,7 +364,7 @@ class TachiyomiManifestParser {
                 val stylesStart = readUInt32(bytes, offset + 24)
 
                 // Read string offsets
-                val stringOffsets = (0 until stringCount).map { i ->
+                val stringOffsets = (0 until stringCount.toInt()).map { i ->
                     readUInt32(bytes, offset + 28 + i * 4)
                 }
